Move MemoryInformation to com.ibm.ws.kernel.boot.core because otherwise it can't be run from a FAT test from kernel.service since that tries to bring in Tr which causes a NoClassDefFoundError
