+++ dev/com.ibm.ws.kernel.boot.core/src/com/ibm/ws/kernel/util/MemoryInformation.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.kernel.service.util;
+package com.ibm.ws.kernel.util;
 
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
