+++ dev/com.ibm.ws.kernel.boot.core/src/com/ibm/ws/kernel/util/OperatingSystemType.java
@@ -1,0 +1,18 @@
+/*******************************************************************************
+ * Copyright (c) 2018 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.kernel.util;
+
+/**
+ * Describes an operating system type.
+ */
+public enum OperatingSystemType {
+    AIX, HPUX, IBMi, Linux, Mac, Solaris, Unknown, Windows, zOS
+}
