+++ dev/com.ibm.ws.kernel.boot.core/src/com/ibm/ws/kernel/util/package-info.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
+ * Copyright (c) 2013 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,11 +8,11 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.kernel.service.util;
-
 /**
- * Describes an operating system type.
+ * @version 1.0
  */
-public enum OperatingSystemType {
-    AIX, HPUX, IBMi, Linux, Mac, Solaris, Unknown, Windows, zOS
-}
+@org.osgi.annotation.versioning.Version("1.0")
+@TraceOptions(messageBundle = "com.ibm.ws.kernel.boot.resources.LauncherMessages")
+package com.ibm.ws.kernel.util;
+
+import com.ibm.websphere.ras.annotation.TraceOptions;
