+++ dev/com.ibm.ws.kernel.service/bnd.bnd
@@ -52,5 +52,3 @@ instrument.classesExcludes: com/ibm/ws/kernel/service/utils/resources/*.class
 	com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
 	org.eclipse.osgi;version=latest
 	
--testpath: \
-  ../build.sharedResources/lib/junit/old/junit.jar;version=file
